refactor(provider): split auth runtime shared core and provider backends (#1684)
