+++ crates/tau-provider/src/auth_commands_runtime.rs
@@ -4,16 +4,24 @@
 //! output into structured records. Command failures and malformed output are
 //! surfaced as typed errors for login/status/logout flows.
 
-use std::process::Command;
-use std::time::Duration;
-
-use anyhow::{bail, Context, Result};
+use anyhow::{bail, Result};
 use serde::Serialize;
 use tau_ai::Provider;
 use tau_core::current_unix_timestamp;
-use wait_timeout::ChildExt;
 
 use crate::*;
+
+mod anthropic_backend;
+mod google_backend;
+mod openai_backend;
+mod shared_runtime_core;
+
+use anthropic_backend::execute_anthropic_login_backend_ready;
+use google_backend::execute_google_login_backend_ready;
+use openai_backend::execute_openai_login_backend_ready;
+use shared_runtime_core::{
+    build_auth_login_launch_spec, collect_non_empty_secrets, redact_known_secrets,
+};
 
 /// Public `const` `AUTH_USAGE` in `tau-provider`.
 ///
@@ -723,138 +731,6 @@ pub fn parse_auth_revoked_filter(token: &str) -> Result<AuthRevokedFilter> {
     }
 }
 
-fn collect_non_empty_secrets(candidates: Vec<(&'static str, Option<String>)>) -> Vec<String> {
-    candidates
-        .into_iter()
-        .filter_map(|(_source, value)| value)
-        .map(|value| value.trim().to_string())
-        .filter(|value| !value.is_empty())
-        .collect()
-}
-
-fn redact_known_secrets(text: String, secrets: &[String]) -> String {
-    let mut redacted = text;
-    for secret in secrets {
-        if !secret.is_empty() {
-            redacted = redacted.replace(secret, "<redacted>");
-        }
-    }
-    redacted
-}
-
-const AUTH_LOGIN_LAUNCH_TIMEOUT_MS: u64 = 300_000;
-
-struct AuthLoginLaunchSpec {
-    executable: String,
-    args: Vec<String>,
-    timeout_ms: u64,
-}
-
-struct AuthLoginLaunchResult {
-    command: String,
-}
-
-fn shell_quote_token(token: &str) -> String {
-    if token
-        .chars()
-        .all(|ch| ch.is_ascii_alphanumeric() || matches!(ch, '-' | '_' | '.' | '/' | ':'))
-    {
-        return token.to_string();
-    }
-    format!("'{}'", token.replace('\'', "'\"'\"'"))
-}
-
-fn render_launch_command(executable: &str, args: &[String]) -> String {
-    let mut parts = Vec::with_capacity(args.len().saturating_add(1));
-    parts.push(shell_quote_token(executable));
-    parts.extend(args.iter().map(|arg| shell_quote_token(arg)));
-    parts.join(" ")
-}
-
-fn execute_auth_login_launch(spec: &AuthLoginLaunchSpec) -> Result<AuthLoginLaunchResult> {
-    let mut command = Command::new(spec.executable.trim());
-    command.args(spec.args.iter().map(String::as_str));
-    command.stdin(std::process::Stdio::inherit());
-    command.stdout(std::process::Stdio::inherit());
-    command.stderr(std::process::Stdio::inherit());
-    let command_str = render_launch_command(spec.executable.as_str(), &spec.args);
-
-    let mut child = command
-        .spawn()
-        .with_context(|| format!("failed to spawn login command {}", command_str))?;
-
-    let timeout = Duration::from_millis(spec.timeout_ms.max(1));
-    let status = match child
-        .wait_timeout(timeout)
-        .with_context(|| format!("failed while waiting for login command {}", command_str))?
-    {
-        Some(status) => status,
-        None => {
-            let _ = child.kill();
-            let _ = child.wait();
-            bail!(
-                "login command timed out after {} ms: {}",
-                spec.timeout_ms.max(1),
-                command_str
-            );
-        }
-    };
-    if !status.success() {
-        let code = status
-            .code()
-            .map(|value| value.to_string())
-            .unwrap_or_else(|| "terminated_by_signal".to_string());
-        bail!("login command exited with status {}: {}", code, command_str);
-    }
-
-    Ok(AuthLoginLaunchResult {
-        command: command_str,
-    })
-}
-
-fn build_auth_login_launch_spec(
-    config: &AuthCommandConfig,
-    provider: Provider,
-    mode: ProviderAuthMethod,
-) -> Result<AuthLoginLaunchSpec> {
-    let timeout_ms = AUTH_LOGIN_LAUNCH_TIMEOUT_MS;
-    match (provider, mode) {
-        (
-            Provider::OpenAi,
-            ProviderAuthMethod::OauthToken | ProviderAuthMethod::SessionToken,
-        ) => Ok(AuthLoginLaunchSpec {
-            executable: config.openai_codex_cli.clone(),
-            args: vec!["--login".to_string()],
-            timeout_ms,
-        }),
-        (
-            Provider::Anthropic,
-            ProviderAuthMethod::OauthToken | ProviderAuthMethod::SessionToken,
-        ) => Ok(AuthLoginLaunchSpec {
-            executable: config.anthropic_claude_cli.clone(),
-            args: Vec::new(),
-            timeout_ms,
-        }),
-        (Provider::Google, ProviderAuthMethod::OauthToken) => Ok(AuthLoginLaunchSpec {
-            executable: config.google_gemini_cli.clone(),
-            args: Vec::new(),
-            timeout_ms,
-        }),
-        (Provider::Google, ProviderAuthMethod::Adc) => Ok(AuthLoginLaunchSpec {
-            executable: config.google_gcloud_cli.clone(),
-            args: vec![
-                "auth".to_string(),
-                "application-default".to_string(),
-                "login".to_string(),
-            ],
-            timeout_ms,
-        }),
-        _ => bail!(
-            "--launch is only supported for openai oauth-token/session-token, anthropic oauth-token/session-token, and google oauth-token/adc"
-        ),
-    }
-}
-
 /// Public `fn` `execute_auth_login_command` in `tau-provider`.
 ///
 /// This item is part of the Wave 2 API surface for M23 documentation uplift.
@@ -1185,503 +1061,6 @@ pub fn execute_auth_login_command(
             )
         }
     }
-}
-
-fn execute_google_login_backend_ready(
-    config: &AuthCommandConfig,
-    mode: ProviderAuthMethod,
-    launch: bool,
-    json_output: bool,
-) -> String {
-    if !config.google_gemini_backend {
-        let reason =
-            "google gemini backend is disabled; set --google-gemini-backend=true".to_string();
-        if json_output {
-            return serde_json::json!({
-                "command": "auth.login",
-                "provider": Provider::Google.as_str(),
-                "mode": mode.as_str(),
-                "status": "error",
-                "reason": reason,
-            })
-            .to_string();
-        }
-        return format!(
-            "auth login error: provider={} mode={} launch_requested={} launch_executed=false error={reason}",
-            Provider::Google.as_str(),
-            mode.as_str(),
-            launch
-        );
-    }
-
-    if !is_executable_available(&config.google_gemini_cli) {
-        let reason = format!(
-            "gemini cli executable '{}' is not available",
-            config.google_gemini_cli
-        );
-        if json_output {
-            return serde_json::json!({
-                "command": "auth.login",
-                "provider": Provider::Google.as_str(),
-                "mode": mode.as_str(),
-                "status": "error",
-                "reason": reason,
-            })
-            .to_string();
-        }
-        return format!(
-            "auth login error: provider={} mode={} launch_requested={} launch_executed=false error={reason}",
-            Provider::Google.as_str(),
-            mode.as_str(),
-            launch
-        );
-    }
-
-    if mode == ProviderAuthMethod::Adc && !is_executable_available(&config.google_gcloud_cli) {
-        let reason = format!(
-            "gcloud executable '{}' is not available",
-            config.google_gcloud_cli
-        );
-        if json_output {
-            return serde_json::json!({
-                "command": "auth.login",
-                "provider": Provider::Google.as_str(),
-                "mode": mode.as_str(),
-                "status": "error",
-                "reason": reason,
-                "launch_requested": launch,
-                "launch_executed": false,
-            })
-            .to_string();
-        }
-        return format!(
-            "auth login error: provider={} mode={} launch_requested={} launch_executed=false error={reason}",
-            Provider::Google.as_str(),
-            mode.as_str(),
-            launch
-        );
-    }
-
-    let action = if mode == ProviderAuthMethod::Adc {
-        "run gcloud auth application-default login and set GOOGLE_CLOUD_PROJECT/GOOGLE_CLOUD_LOCATION"
-    } else {
-        "run gemini and select Login with Google"
-    };
-    let launch_spec = match build_auth_login_launch_spec(config, Provider::Google, mode) {
-        Ok(spec) => spec,
-        Err(error) => {
-            if json_output {
-                return serde_json::json!({
-                    "command": "auth.login",
-                    "provider": Provider::Google.as_str(),
-                    "mode": mode.as_str(),
-                    "status": "error",
-                    "reason": error.to_string(),
-                    "launch_requested": launch,
-                    "launch_executed": false,
-                })
-                .to_string();
-            }
-            return format!(
-                "auth login error: provider={} mode={} launch_requested={} launch_executed=false error={error}",
-                Provider::Google.as_str(),
-                mode.as_str(),
-                launch
-            );
-        }
-    };
-    let launch_command = render_launch_command(&launch_spec.executable, &launch_spec.args);
-
-    if launch {
-        return match execute_auth_login_launch(&launch_spec) {
-            Ok(result) => {
-                if json_output {
-                    serde_json::json!({
-                        "command": "auth.login",
-                        "provider": Provider::Google.as_str(),
-                        "mode": mode.as_str(),
-                        "status": "launched",
-                        "source": "gemini_cli",
-                        "backend_cli": config.google_gemini_cli,
-                        "persisted": false,
-                        "action": action,
-                        "launch_requested": true,
-                        "launch_executed": true,
-                        "launch_command": result.command,
-                    })
-                    .to_string()
-                } else {
-                    format!(
-                        "auth login: provider={} mode={} status=launched source=gemini_cli backend_cli={} persisted=false action={} launch_requested=true launch_executed=true launch_command={}",
-                        Provider::Google.as_str(),
-                        mode.as_str(),
-                        config.google_gemini_cli,
-                        action,
-                        result.command
-                    )
-                }
-            }
-            Err(error) => {
-                if json_output {
-                    serde_json::json!({
-                        "command": "auth.login",
-                        "provider": Provider::Google.as_str(),
-                        "mode": mode.as_str(),
-                        "status": "error",
-                        "reason": error.to_string(),
-                        "launch_requested": true,
-                        "launch_executed": false,
-                        "launch_command": launch_command,
-                    })
-                    .to_string()
-                } else {
-                    format!(
-                        "auth login error: provider={} mode={} launch_requested=true launch_executed=false launch_command={} error={error}",
-                        Provider::Google.as_str(),
-                        mode.as_str(),
-                        launch_command
-                    )
-                }
-            }
-        };
-    }
-
-    if json_output {
-        return serde_json::json!({
-            "command": "auth.login",
-            "provider": Provider::Google.as_str(),
-            "mode": mode.as_str(),
-            "status": "ready",
-            "source": "gemini_cli",
-            "backend_cli": config.google_gemini_cli,
-            "persisted": false,
-            "action": action,
-            "launch_requested": false,
-            "launch_executed": false,
-            "launch_command": launch_command,
-        })
-        .to_string();
-    }
-    format!(
-        "auth login: provider={} mode={} status=ready source=gemini_cli backend_cli={} persisted=false action={} launch_requested=false launch_executed=false launch_command={}",
-        Provider::Google.as_str(),
-        mode.as_str(),
-        config.google_gemini_cli,
-        action,
-        launch_command
-    )
-}
-
-fn execute_openai_login_backend_ready(
-    config: &AuthCommandConfig,
-    mode: ProviderAuthMethod,
-    launch: bool,
-    json_output: bool,
-) -> String {
-    if !config.openai_codex_backend {
-        let reason =
-            "openai codex backend is disabled; set --openai-codex-backend=true".to_string();
-        if json_output {
-            return serde_json::json!({
-                "command": "auth.login",
-                "provider": Provider::OpenAi.as_str(),
-                "mode": mode.as_str(),
-                "status": "error",
-                "reason": reason,
-            })
-            .to_string();
-        }
-        return format!(
-            "auth login error: provider={} mode={} launch_requested={} launch_executed=false error={reason}",
-            Provider::OpenAi.as_str(),
-            mode.as_str(),
-            launch
-        );
-    }
-
-    if !is_executable_available(&config.openai_codex_cli) {
-        let reason = format!(
-            "codex cli executable '{}' is not available",
-            config.openai_codex_cli
-        );
-        if json_output {
-            return serde_json::json!({
-                "command": "auth.login",
-                "provider": Provider::OpenAi.as_str(),
-                "mode": mode.as_str(),
-                "status": "error",
-                "reason": reason,
-            })
-            .to_string();
-        }
-        return format!(
-            "auth login error: provider={} mode={} launch_requested={} launch_executed=false error={reason}",
-            Provider::OpenAi.as_str(),
-            mode.as_str(),
-            launch
-        );
-    }
-
-    let action = "run codex --login";
-    let launch_spec = match build_auth_login_launch_spec(config, Provider::OpenAi, mode) {
-        Ok(spec) => spec,
-        Err(error) => {
-            if json_output {
-                return serde_json::json!({
-                    "command": "auth.login",
-                    "provider": Provider::OpenAi.as_str(),
-                    "mode": mode.as_str(),
-                    "status": "error",
-                    "reason": error.to_string(),
-                    "launch_requested": launch,
-                    "launch_executed": false,
-                })
-                .to_string();
-            }
-            return format!(
-                "auth login error: provider={} mode={} launch_requested={} launch_executed=false error={error}",
-                Provider::OpenAi.as_str(),
-                mode.as_str(),
-                launch
-            );
-        }
-    };
-    let launch_command = render_launch_command(&launch_spec.executable, &launch_spec.args);
-
-    if launch {
-        return match execute_auth_login_launch(&launch_spec) {
-            Ok(result) => {
-                if json_output {
-                    serde_json::json!({
-                        "command": "auth.login",
-                        "provider": Provider::OpenAi.as_str(),
-                        "mode": mode.as_str(),
-                        "status": "launched",
-                        "source": "codex_cli",
-                        "backend_cli": config.openai_codex_cli,
-                        "persisted": false,
-                        "action": action,
-                        "launch_requested": true,
-                        "launch_executed": true,
-                        "launch_command": result.command,
-                    })
-                    .to_string()
-                } else {
-                    format!(
-                        "auth login: provider={} mode={} status=launched source=codex_cli backend_cli={} persisted=false action={} launch_requested=true launch_executed=true launch_command={}",
-                        Provider::OpenAi.as_str(),
-                        mode.as_str(),
-                        config.openai_codex_cli,
-                        action,
-                        result.command
-                    )
-                }
-            }
-            Err(error) => {
-                if json_output {
-                    serde_json::json!({
-                        "command": "auth.login",
-                        "provider": Provider::OpenAi.as_str(),
-                        "mode": mode.as_str(),
-                        "status": "error",
-                        "reason": error.to_string(),
-                        "launch_requested": true,
-                        "launch_executed": false,
-                        "launch_command": launch_command,
-                    })
-                    .to_string()
-                } else {
-                    format!(
-                        "auth login error: provider={} mode={} launch_requested=true launch_executed=false launch_command={} error={error}",
-                        Provider::OpenAi.as_str(),
-                        mode.as_str(),
-                        launch_command
-                    )
-                }
-            }
-        };
-    }
-
-    if json_output {
-        return serde_json::json!({
-            "command": "auth.login",
-            "provider": Provider::OpenAi.as_str(),
-            "mode": mode.as_str(),
-            "status": "ready",
-            "source": "codex_cli",
-            "backend_cli": config.openai_codex_cli,
-            "persisted": false,
-            "action": action,
-            "launch_requested": false,
-            "launch_executed": false,
-            "launch_command": launch_command,
-        })
-        .to_string();
-    }
-    format!(
-        "auth login: provider={} mode={} status=ready source=codex_cli backend_cli={} persisted=false action={} launch_requested=false launch_executed=false launch_command={}",
-        Provider::OpenAi.as_str(),
-        mode.as_str(),
-        config.openai_codex_cli,
-        action,
-        launch_command
-    )
-}
-
-fn execute_anthropic_login_backend_ready(
-    config: &AuthCommandConfig,
-    mode: ProviderAuthMethod,
-    launch: bool,
-    json_output: bool,
-) -> String {
-    if !config.anthropic_claude_backend {
-        let reason =
-            "anthropic claude backend is disabled; set --anthropic-claude-backend=true".to_string();
-        if json_output {
-            return serde_json::json!({
-                "command": "auth.login",
-                "provider": Provider::Anthropic.as_str(),
-                "mode": mode.as_str(),
-                "status": "error",
-                "reason": reason,
-            })
-            .to_string();
-        }
-        return format!(
-            "auth login error: provider={} mode={} launch_requested={} launch_executed=false error={reason}",
-            Provider::Anthropic.as_str(),
-            mode.as_str(),
-            launch
-        );
-    }
-
-    if !is_executable_available(&config.anthropic_claude_cli) {
-        let reason = format!(
-            "claude cli executable '{}' is not available",
-            config.anthropic_claude_cli
-        );
-        if json_output {
-            return serde_json::json!({
-                "command": "auth.login",
-                "provider": Provider::Anthropic.as_str(),
-                "mode": mode.as_str(),
-                "status": "error",
-                "reason": reason,
-            })
-            .to_string();
-        }
-        return format!(
-            "auth login error: provider={} mode={} launch_requested={} launch_executed=false error={reason}",
-            Provider::Anthropic.as_str(),
-            mode.as_str(),
-            launch
-        );
-    }
-
-    let action = "run claude, then enter /login in the Claude prompt";
-    let launch_spec = match build_auth_login_launch_spec(config, Provider::Anthropic, mode) {
-        Ok(spec) => spec,
-        Err(error) => {
-            if json_output {
-                return serde_json::json!({
-                    "command": "auth.login",
-                    "provider": Provider::Anthropic.as_str(),
-                    "mode": mode.as_str(),
-                    "status": "error",
-                    "reason": error.to_string(),
-                    "launch_requested": launch,
-                    "launch_executed": false,
-                })
-                .to_string();
-            }
-            return format!(
-                "auth login error: provider={} mode={} launch_requested={} launch_executed=false error={error}",
-                Provider::Anthropic.as_str(),
-                mode.as_str(),
-                launch
-            );
-        }
-    };
-    let launch_command = render_launch_command(&launch_spec.executable, &launch_spec.args);
-
-    if launch {
-        return match execute_auth_login_launch(&launch_spec) {
-            Ok(result) => {
-                if json_output {
-                    serde_json::json!({
-                        "command": "auth.login",
-                        "provider": Provider::Anthropic.as_str(),
-                        "mode": mode.as_str(),
-                        "status": "launched",
-                        "source": "claude_cli",
-                        "backend_cli": config.anthropic_claude_cli,
-                        "persisted": false,
-                        "action": action,
-                        "launch_requested": true,
-                        "launch_executed": true,
-                        "launch_command": result.command,
-                    })
-                    .to_string()
-                } else {
-                    format!(
-                        "auth login: provider={} mode={} status=launched source=claude_cli backend_cli={} persisted=false action={} launch_requested=true launch_executed=true launch_command={}",
-                        Provider::Anthropic.as_str(),
-                        mode.as_str(),
-                        config.anthropic_claude_cli,
-                        action,
-                        result.command
-                    )
-                }
-            }
-            Err(error) => {
-                if json_output {
-                    serde_json::json!({
-                        "command": "auth.login",
-                        "provider": Provider::Anthropic.as_str(),
-                        "mode": mode.as_str(),
-                        "status": "error",
-                        "reason": error.to_string(),
-                        "launch_requested": true,
-                        "launch_executed": false,
-                        "launch_command": launch_command,
-                    })
-                    .to_string()
-                } else {
-                    format!(
-                        "auth login error: provider={} mode={} launch_requested=true launch_executed=false launch_command={} error={error}",
-                        Provider::Anthropic.as_str(),
-                        mode.as_str(),
-                        launch_command
-                    )
-                }
-            }
-        };
-    }
-
-    if json_output {
-        return serde_json::json!({
-            "command": "auth.login",
-            "provider": Provider::Anthropic.as_str(),
-            "mode": mode.as_str(),
-            "status": "ready",
-            "source": "claude_cli",
-            "backend_cli": config.anthropic_claude_cli,
-            "persisted": false,
-            "action": action,
-            "launch_requested": false,
-            "launch_executed": false,
-            "launch_command": launch_command,
-        })
-        .to_string();
-    }
-    format!(
-        "auth login: provider={} mode={} status=ready source=claude_cli backend_cli={} persisted=false action={} launch_requested=false launch_executed=false launch_command={}",
-        Provider::Anthropic.as_str(),
-        mode.as_str(),
-        config.anthropic_claude_cli,
-        action,
-        launch_command
-    )
 }
 
 fn auth_reauth_status_label(row: &AuthStatusRow) -> &'static str {
